Added IsBusy functionality to CustomSwitch control
Two new files, `IsBusyTestPage.xaml.cs` and `IsBusyTestPage.xaml`, have been added to `App.csproj`. The `CustomSwitch` control in `IosSwitch.xaml` and `CustomSwitch.xaml` now has a new `IsBusy` property, indicating whether the switch is busy, and an `IsBusyContent` property, which displays an `ActivityIndicator` when the switch is busy. A new button has been added to `MainPage.xaml` that navigates to the `IsBusyTestPage`. The `IsBusyTestPage.xaml` contains two `IosSwitch` controls bound to the `IsToggled` property of a `DynamicToggle` switch.