Improve nuget release security (#28)
Updated Release.yml to use GitHub OIDC for obtaining a short-lived NuGet API key, replacing the use of long-lived secrets.
Renamed the `build` job to `build-and-publish` and added a `permissions` section to enable OIDC token issuance.
Introduced a `NuGet login` step to authenticate with NuGet using OIDC, and updated the `Publish` step to use the temporary API key.